requirements.txt
fastapi
uvicorn
fpdf2